refactor(AbstractUserController): Extract AbstractUserController from Otp/Admin/Api-UserController.
