+++ src/main/java/org/opentripplanner/middleware/controllers/api/ApiUserController.java
@@ -1,106 +1,31 @@
 package org.opentripplanner.middleware.controllers.api;
 
-import com.auth0.exception.Auth0Exception;
-import com.auth0.json.mgmt.users.User;
-import com.beerboy.ss.ApiEndpoint;
-import org.eclipse.jetty.http.HttpStatus;
-import org.opentripplanner.middleware.auth.Auth0Connection;
 import org.opentripplanner.middleware.auth.Auth0UserProfile;
 import org.opentripplanner.middleware.models.ApiUser;
 import org.opentripplanner.middleware.persistence.Persistence;
-import org.opentripplanner.middleware.utils.JsonUtils;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 import spark.Request;
-import spark.Response;
-
-import static com.beerboy.ss.descriptor.MethodDescriptor.path;
-import static org.opentripplanner.middleware.auth.Auth0Users.createNewAuth0User;
-import static org.opentripplanner.middleware.auth.Auth0Users.deleteAuth0User;
-import static org.opentripplanner.middleware.auth.Auth0Users.updateAuthFieldsForUser;
-import static org.opentripplanner.middleware.auth.Auth0Users.validateExistingUser;
-import static org.opentripplanner.middleware.utils.JsonUtils.logMessageAndHalt;
 
 /**
- * Implementation of the {@link ApiController} abstract class for managing API users. This controller connects with Auth0
- * services using the hooks provided by {@link ApiController}.
+ * Implementation of the {@link AbstractUserController} for {@link ApiUser}.
  */
-public class ApiUserController extends ApiController<ApiUser> {
-    private static final Logger LOG = LoggerFactory.getLogger(ApiUserController.class);
-    static final String NO_USER_WITH_AUTH0_ID_MESSAGE = "No user with auth0UserID=%s found.";
-    private static final String TOKEN_PATH = "/fromtoken";
-
+public class ApiUserController extends AbstractUserController<ApiUser> {
     public ApiUserController(String apiPrefix){
         super(apiPrefix, Persistence.apiUsers, "secure/application");
     }
 
     @Override
-    protected void buildEndpoint(ApiEndpoint baseEndpoint) {
-        LOG.info("Registering path {}.", ROOT_ROUTE + TOKEN_PATH);
-
-        // Add the user token route BEFORE the regular CRUD methods
-        // (otherwise, /fromtoken requests would be considered
-        // by spark as 'GET user with id "fromtoken"', which we don't want).
-        ApiEndpoint modifiedEndpoint = baseEndpoint
-            // Get user from token.
-            .get(path(ROOT_ROUTE + TOKEN_PATH)
-                    .withDescription("Retrieves an ApiUser entity using an Auth0 access token passed in an Authorization header.")
-                    .withResponseType(persistence.clazz),
-                this::getUserFromRequest, JsonUtils::toJson
-            )
-
-            // Options response for CORS for the token path
-            .options(path(TOKEN_PATH), (req, res) -> "");
-
-        // Add the regular CRUD methods after defining the /fromtoken route.
-        super.buildEndpoint(modifiedEndpoint);
+    protected ApiUser getUserProfile(Auth0UserProfile profile) {
+        return profile.apiUser;
     }
 
     /**
-     * HTTP endpoint to get the {@link ApiUser} entity, if it exists, from an {@link Auth0UserProfile} attribute
-     * available from a {@link Request} (this is the case for '/api/secure/' endpoints).
-     */
-    private ApiUser getUserFromRequest(Request req, Response res) {
-        Auth0UserProfile profile = Auth0Connection.getUserFromRequest(req);
-        ApiUser user = profile.apiUser;
-
-        // If the ApiUser object is null, it is most likely because it was not created yet,
-        // for instance, for users who just created an Auth0 login and have an Auth0UserProfile
-        // but have not completed the account setup form yet.
-        // For those users, the ApiUser profile would be 404 not found (as opposed to 403 forbidden).
-        if (user == null) {
-            logMessageAndHalt(req, HttpStatus.NOT_FOUND_404, String.format(NO_USER_WITH_AUTH0_ID_MESSAGE, profile.auth0UserId), null);
-        }
-        return user;
-    }
-
-    /**
-     * Before creating/storing a user in MongoDB, create the user in Auth0 and update the {@link ApiUser#auth0UserId}
-     * with the value from Auth0.
+     * Before creating/storing an API user in MongoDB,
+     * add the API keys for this user.
      */
     @Override
     ApiUser preCreateHook(ApiUser user, Request req) {
-        User auth0UserProfile = createNewAuth0User(user, req, this.persistence);
-        return updateAuthFieldsForUser(user, auth0UserProfile);
-    }
+        // TODO: Add the API keys for this user.
 
-    @Override
-    ApiUser preUpdateHook(ApiUser user, ApiUser preExistingUser, Request req) {
-        validateExistingUser(user, preExistingUser, req, this.persistence);
-        return user;
-    }
-
-    /**
-     * Before deleting the user in MongoDB, attempt to delete the user in Auth0.
-     */
-    @Override
-    boolean preDeleteHook(ApiUser user, Request req) {
-        try {
-            deleteAuth0User(user.auth0UserId);
-        } catch (Auth0Exception e) {
-            logMessageAndHalt(req, 500, "Error deleting user.", e);
-            return false;
-        }
-        return true;
+        return super.preCreateHook(user, req);
     }
 }
